Remove UploadWorkManager
