+++ engine/src/main/java/com/google/android/fhir/sync/Config.kt
@@ -40,7 +40,6 @@ val defaultRetryConfiguration =
 object SyncDataParams {
   const val SORT_KEY = "_sort"
   const val LAST_UPDATED_KEY = "_lastUpdated"
-  const val ADDRESS_COUNTRY_KEY = "address-country"
   const val SUMMARY_KEY = "_summary"
   const val SUMMARY_COUNT_VALUE = "count"
 }
