+++ engine/src/main/java/com/google/android/fhir/sync/FhirSynchronizer.kt
@@ -19,6 +19,8 @@ package com.google.android.fhir.sync
 import android.content.Context
 import com.google.android.fhir.DatastoreUtil
 import com.google.android.fhir.FhirEngine
+import com.google.android.fhir.sync.download.DownloadState
+import com.google.android.fhir.sync.download.Downloader
 import com.google.android.fhir.sync.upload.UploadState
 import com.google.android.fhir.sync.upload.Uploader
 import java.time.OffsetDateTime
