+++ engine/src/main/java/com/google/android/fhir/sync/download/DownloaderImpl.kt
@@ -19,9 +19,7 @@ package com.google.android.fhir.sync.download
 import com.google.android.fhir.sync.BundleDownloadRequest
 import com.google.android.fhir.sync.DataSource
 import com.google.android.fhir.sync.DownloadRequest
-import com.google.android.fhir.sync.DownloadState
 import com.google.android.fhir.sync.DownloadWorkManager
-import com.google.android.fhir.sync.Downloader
 import com.google.android.fhir.sync.ResourceSyncException
 import com.google.android.fhir.sync.UrlDownloadRequest
 import kotlinx.coroutines.flow.Flow
