+++ engine/src/main/java/com/google/android/fhir/sync/upload/Uploader.kt
@@ -18,19 +18,83 @@ package com.google.android.fhir.sync.upload
 
 import com.google.android.fhir.LocalChange
 import com.google.android.fhir.LocalChangeToken
+import com.google.android.fhir.sync.DataSource
 import com.google.android.fhir.sync.ResourceSyncException
+import com.google.android.fhir.sync.upload.patch.PerResourcePatchGenerator
+import com.google.android.fhir.sync.upload.request.TransactionBundleGenerator
 import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.flow
+import org.hl7.fhir.exceptions.FHIRException
+import org.hl7.fhir.r4.model.Bundle
+import org.hl7.fhir.r4.model.OperationOutcome
 import org.hl7.fhir.r4.model.Resource
+import org.hl7.fhir.r4.model.ResourceType
+import timber.log.Timber
 
-/** Module for uploading local changes to a [DataSource]. */
-internal interface Uploader {
+/**
+ * Implementation of the [Uploader]. It orchestrates the pre processing of [LocalChange] and
+ * constructing appropriate upload requests via [UploadWorkManager] and uploading of requests via
+ * [DataSource]. [Uploader] clients should call upload and listen to the various states emitted by
+ * [UploadWorkManager] as [UploadState].
+ */
+internal class Uploader(
+  private val dataSource: DataSource,
+) {
+  private val patchGenerator = PerResourcePatchGenerator
+  private val requestGenerator = TransactionBundleGenerator.getDefault()
 
   /**
    * Uploads the local changes to the [DataSource]. Particular implementations should take care of
    * transforming the [LocalChange]s to particular network operations. If [ProgressCallback] is
    * provided it also reports the intermediate progress
    */
-  suspend fun upload(localChanges: List<LocalChange>): Flow<UploadState>
+  suspend fun upload(localChanges: List<LocalChange>): Flow<UploadState> = flow {
+    val patches = patchGenerator.generate(localChanges)
+    val requests = requestGenerator.generateUploadRequests(patches)
+    val token = LocalChangeToken(localChanges.flatMap { it.token.ids })
+    val total = requests.size
+    emit(UploadState.Started(total))
+    requests.forEachIndexed { index, uploadRequest ->
+      try {
+        val response = dataSource.upload(uploadRequest)
+        emit(
+          getUploadResult(uploadRequest.resource.resourceType, response, token, total, index + 1)
+        )
+      } catch (e: Exception) {
+        Timber.e(e)
+        emit(UploadState.Failure(ResourceSyncException(ResourceType.Bundle, e)))
+      }
+    }
+  }
+
+  private fun getUploadResult(
+    requestResourceType: ResourceType,
+    response: Resource,
+    localChangeToken: LocalChangeToken,
+    total: Int,
+    completed: Int
+  ) =
+    when {
+      response is Bundle && response.type == Bundle.BundleType.TRANSACTIONRESPONSE -> {
+        UploadState.Success(localChangeToken, response, total, completed)
+      }
+      response is OperationOutcome && response.issue.isNotEmpty() -> {
+        UploadState.Failure(
+          ResourceSyncException(
+            requestResourceType,
+            FHIRException(response.issueFirstRep.diagnostics)
+          )
+        )
+      }
+      else -> {
+        UploadState.Failure(
+          ResourceSyncException(
+            requestResourceType,
+            FHIRException("Unknown response for ${response.resourceType}")
+          )
+        )
+      }
+    }
 }
 
 internal sealed class UploadState {
